+++ helpers/getContacts.js
@@ -1,0 +1,14 @@
+const contacts = require('../models/contacts')
+
+const getContacts = async (req, res, next) => {
+  try {
+    const result = await contacts.listContacts()
+    res.json(result)
+  } catch (error) {
+    next(error)
+    
+  }
+
+}
+
+module.exports = getContacts
