+++ helpers/index.js
@@ -1,5 +1,15 @@
 const HttpError = require("./HttpError")
+const getContacts = require("./getContacts")
+const deleteById = require("./deleteById")
+const getById = require("./getById")
+const postContact = require("./postContact")
+const putContact = require("./putContact")
 
 module.exports = {
   HttpError,
+  getContacts,
+  deleteById,
+  getById,
+  postContact,
+  putContact,
 }
