+++ helpers/validation.js
@@ -1,0 +1,16 @@
+const Joi = require("joi")
+
+const addShema = Joi.object({
+  name: Joi.string().required().messages({
+    "any.required": "missing required name field"
+  }),
+  email: Joi.string().required().messages({
+    "any.required": "missing required email field"
+  }),
+  phone: Joi.string().required().messages({
+    "any.required": "missing required phone field"
+  }),
+
+})
+
+module.exports = addShema
