+++ models/contacts.json
@@ -1,7 +1,7 @@
 [
   {
     "id": "AeHIrLTr6JkxGE6SN-0Rw",
-    "name": "Allen Raymond",
+    "name": "Allen Raymond III",
     "email": "nulla.ante@vestibul.co.uk",
     "phone": "(992) 914-3792"
   },
@@ -54,15 +54,9 @@
     "phone": "(704) 398-7993"
   },
   {
-    "id": "rsKkOQUi80UsgVPCcLZZW",
-    "name": "Alec Howard III",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
-  },
-  {
-    "id": "5fpz2mgwmWwKbLD7-sG-m",
-    "name": "user",
+    "id": "THS4FqOE4zfUWB-V2fuht",
+    "name": "Allen Raymond",
     "email": "nulla.ante@vestibul.co.uk",
-    "phone": "33333433"
+    "phone": "(992) 914-3792"
   }
 ]
